requirements.txt
python-dotenv~=1.0
requests~=2.31
marshmallow~=3.2